+++ ownlang-parser/src/main/java/com/annimon/ownlang/parser/TokenType.java
@@ -7,8 +7,10 @@ package com.annimon.ownlang.parser;
 public enum TokenType {
 
     NUMBER,
+    LONG_NUMBER,
     DECIMAL_NUMBER,
     HEX_NUMBER,
+    HEX_LONG_NUMBER,
     WORD,
     TEXT,